requirements.txt
whisper
llama-cpp-python
torch
textwrap3